Refs 10495. Moving LocatorList_t to eprosima::fastdds::rtps.
