+++ include/fastdds/rtps/common/Locator.h
@@ -368,347 +368,30 @@ inline std::istream& operator >>(
 typedef std::vector<Locator_t>::iterator LocatorListIterator;
 typedef std::vector<Locator_t>::const_iterator LocatorListConstIterator;
 
-/**
- * Provides a Locator's iterator interface that can be used by different Locator's
- * containers
- */
-class LocatorsIterator
-{
-public:
+} // namespace rtps
+} // namespace fastrtps
+} // namespace eprosima
 
-    virtual LocatorsIterator& operator ++() = 0;
-    virtual bool operator ==(
-            const LocatorsIterator& other) const = 0;
-    virtual bool operator !=(
-            const LocatorsIterator& other) const = 0;
-    virtual const Locator_t& operator *() const = 0;
-};
+namespace eprosima {
+namespace fastdds {
+namespace rtps {
 
-/**
- * Adapter class that provides a LocatorsIterator interface from a LocatorListConstIterator
- */
-class Locators : public LocatorsIterator
-{
-public:
+using Locator = eprosima::fastrtps::rtps::Locator_t;
 
-    Locators(
-            const LocatorListConstIterator& it)
-        : it_(it)
-    {
-    }
+} // namespace rtps
+} // namespace fastdds
+} // namespace eprosima
 
-    Locators(
-            const Locators& other)
-        : it_(other.it_)
-    {
-    }
+#include <fastdds/rtps/common/LocatorsIterator.hpp>
+#include <fastdds/rtps/common/LocatorList.hpp>
 
-    LocatorsIterator& operator ++()
-    {
-        ++it_;
-        return *this;
-    }
+namespace eprosima {
+namespace fastrtps {
+namespace rtps {
 
-    bool operator ==(
-            const LocatorsIterator& other) const
-    {
-        return it_ == static_cast<const Locators&>(other).it_;
-    }
-
-    bool operator !=(
-            const LocatorsIterator& other) const
-    {
-        return it_ != static_cast<const Locators&>(other).it_;
-    }
-
-    const Locator_t& operator *() const
-    {
-        return (*it_);
-    }
-
-private:
-
-    LocatorListConstIterator it_;
-};
-
-
-/**
- * Class LocatorList_t, a Locator_t vector that doesn't avoid duplicates.
- * @ingroup COMMON_MODULE
- */
-class LocatorList_t
-{
-public:
-
-    RTPS_DllAPI LocatorList_t()
-    {
-    }
-
-    RTPS_DllAPI ~LocatorList_t()
-    {
-    }
-
-    RTPS_DllAPI LocatorList_t(
-            const LocatorList_t& list)
-        : m_locators(list.m_locators)
-    {
-    }
-
-    RTPS_DllAPI LocatorList_t(
-            LocatorList_t&& list)
-        : m_locators(std::move(list.m_locators))
-    {
-    }
-
-    RTPS_DllAPI LocatorList_t& operator =(
-            const LocatorList_t& list)
-    {
-        m_locators = list.m_locators;
-        return *this;
-    }
-
-    RTPS_DllAPI LocatorList_t& operator =(
-            LocatorList_t&& list)
-    {
-        m_locators = std::move(list.m_locators);
-        return *this;
-    }
-
-    RTPS_DllAPI bool operator ==(
-            const LocatorList_t& locator_list) const
-    {
-        if (locator_list.m_locators.size() == m_locators.size())
-        {
-            bool returnedValue = true;
-
-            for (auto it = locator_list.m_locators.begin(); returnedValue &&
-                    it != locator_list.m_locators.end(); ++it)
-            {
-                returnedValue = false;
-
-                for (auto it2 = m_locators.begin(); !returnedValue && it2 != m_locators.end(); ++it2)
-                {
-                    if (*it == *it2)
-                    {
-                        returnedValue = true;
-                    }
-                }
-            }
-
-            return returnedValue;
-        }
-
-        return false;
-    }
-
-    RTPS_DllAPI LocatorListIterator begin()
-    {
-        return m_locators.begin();
-    }
-
-    RTPS_DllAPI LocatorListIterator end()
-    {
-        return m_locators.end();
-    }
-
-    RTPS_DllAPI LocatorListConstIterator begin() const
-    {
-        return m_locators.begin();
-    }
-
-    RTPS_DllAPI LocatorListConstIterator end() const
-    {
-        return m_locators.end();
-    }
-
-    RTPS_DllAPI size_t size() const
-    {
-        return m_locators.size();
-    }
-
-    RTPS_DllAPI LocatorList_t& assign(
-            const LocatorList_t& list)
-    {
-        if (!(*this == list))
-        {
-            m_locators = list.m_locators;
-        }
-        return *this;
-    }
-
-    RTPS_DllAPI void clear()
-    {
-        return m_locators.clear();
-    }
-
-    RTPS_DllAPI void reserve(
-            size_t num)
-    {
-        return m_locators.reserve(num);
-    }
-
-    RTPS_DllAPI void resize(
-            size_t num)
-    {
-        return m_locators.resize(num);
-    }
-
-    RTPS_DllAPI void push_back(
-            const Locator_t& loc)
-    {
-        bool already = false;
-        for (LocatorListIterator it = this->begin(); it != this->end(); ++it)
-        {
-            if (loc == *it)
-            {
-                already = true;
-                break;
-            }
-        }
-        if (!already)
-        {
-            m_locators.push_back(loc);
-        }
-    }
-
-    RTPS_DllAPI void push_back(
-            const LocatorList_t& locList)
-    {
-        for (auto it = locList.m_locators.begin(); it != locList.m_locators.end(); ++it)
-        {
-            this->push_back(*it);
-        }
-    }
-
-    RTPS_DllAPI bool empty() const
-    {
-        return m_locators.empty();
-    }
-
-    RTPS_DllAPI void erase(
-            const Locator_t& loc)
-    {
-        auto it = std::find(m_locators.begin(), m_locators.end(), loc);
-        if (it != m_locators.end())
-        {
-            m_locators.erase(it);
-        }
-    }
-
-    FASTDDS_DEPRECATED_UNTIL(3, "eprosima::fastrtps::rtps::LocatorList_t::contains(const Locator_t&)",
-            "Unused method.")
-    RTPS_DllAPI bool contains(
-            const Locator_t& loc)
-    {
-        for (LocatorListIterator it = this->begin(); it != this->end(); ++it)
-        {
-            if (IsAddressDefined(*it))
-            {
-                if (loc == *it)
-                {
-                    return true;
-                }
-            }
-            else
-            {
-                if (loc.kind == (*it).kind && loc.port == (*it).port)
-                {
-                    return true;
-                }
-            }
-        }
-
-        return false;
-    }
-
-    RTPS_DllAPI bool isValid() const
-    {
-        for (LocatorListConstIterator it = this->begin(); it != this->end(); ++it)
-        {
-            if (!IsLocatorValid(*it))
-            {
-                return false;
-            }
-        }
-        return true;
-    }
-
-    RTPS_DllAPI void swap(
-            LocatorList_t& locatorList)
-    {
-        this->m_locators.swap(locatorList.m_locators);
-    }
-
-    friend std::ostream& operator <<(
-            std::ostream& output,
-            const LocatorList_t& loc);
-
-private:
-
-    std::vector<Locator_t> m_locators;
-};
-
-inline std::ostream& operator <<(
-        std::ostream& output,
-        const LocatorList_t& locList)
-{
-    output << "[";
-    if (!locList.empty())
-    {
-        output << *(locList.begin());
-        for (auto it = locList.m_locators.begin() + 1; it != locList.m_locators.end(); ++it)
-        {
-            output << "," << *it;
-        }
-    }
-    else
-    {
-        output << "_";
-    }
-    output << "]";
-    return output;
-}
-
-inline std::istream& operator >>(
-        std::istream& input,
-        LocatorList_t& locList)
-{
-    std::istream::sentry s(input);
-    locList = LocatorList_t();
-
-    if (s)
-    {
-        char punct;
-        Locator_t loc;
-        std::ios_base::iostate excp_mask = input.exceptions();
-
-        try
-        {
-            input.exceptions(excp_mask | std::ios_base::failbit | std::ios_base::badbit);
-
-            // Get [
-            input >> punct;
-
-            while (punct != ']')
-            {
-                input >> loc >> punct;
-                if (loc.kind != LOCATOR_KIND_INVALID)
-                {
-                    locList.push_back(loc);
-                }
-            }
-        }
-        catch (std::ios_base::failure& )
-        {
-            locList.clear();
-            logWarning(LOCATOR_LIST, "Error deserializing LocatorList");
-        }
-
-        input.exceptions(excp_mask);
-    }
-
-    return input;
-}
+using LocatorsIterator = eprosima::fastdds::rtps::LocatorsIterator;
+using Locators = eprosima::fastdds::rtps::Locators;
+using LocatorList_t = eprosima::fastdds::rtps::LocatorList;
 
 } // namespace rtps
 } // namespace fastrtps
